Bugs squashed & tidy ups

+++ CargoMonitor/CommodityPurchasedEvent.cs
@@ -25,9 +25,9 @@ namespace EddiCargoMonitor
 
         public Commodity commodity { get; private set; }
         public int amount { get; private set; }
-        public long price { get; private set; }
+        public int price { get; private set; }
 
-        public CommodityPurchasedEvent(DateTime timestamp, Commodity commodity, int amount, long price) : base(timestamp, NAME)
+        public CommodityPurchasedEvent(DateTime timestamp, Commodity commodity, int amount, int price) : base(timestamp, NAME)
         {
             this.commodity = commodity;
             this.amount = amount;
